refactor: better organized project structure
