+++ src/routes/products.router.js
@@ -1,5 +1,5 @@
 import { Router } from 'express';
-import { ProductManager } from './../ProductManager.js';
+import { ProductManager } from '../controller/ProductManager.js';
 
 const router = Router();
